keeping JWT filter in a separate package
