+++ dress-design-backend/src/main/java/com/fashion/filter/JWTTokenAuthorizationOncePerRequestFilter.java
@@ -1,4 +1,4 @@
-package com.fashion.auth;
+package com.fashion.filter;
 
 import java.io.IOException;
